+++ .story/stories/index.js
@@ -1,4 +1,7 @@
+import Timeout from './Timeout.story';
 import UIFinTileStories from './UIFinTile.story';
 import EditableTableStories from './EditableTable.story';
 import CarouselStories from './Carousel.story';
 import PieChart from './PieChart.story';
+import Suspense from './Suspense.story';
+import ReactMemo from './React.memo.story';