+++ server/wacai/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  TOKEN_PATH: './server/wacai/token.txt',
+  LOG_PATH: './log/wacai/'
+};