+++ server/wacai/db/config.js
@@ -1,0 +1,5 @@
+// SQLITE db file path
+module.exports = {
+  DB_FILE_PATH: './wacai.db',
+  TABLE_NAME: 'FIN',
+};