+++ server/wacai/db/dao.js
@@ -1,0 +1,66 @@
+const sqlite3 = require('sqlite3').verbose();
+const {DB_FILE_PATH, TABLE_NAME} = require('./config');
+const {logDBError, logDBSuccess} = require('./util');
+
+const createDBConnection = () => {
+  let db = new sqlite3.Database(DB_FILE_PATH, sqlite3.OPEN_READWRITE, (err) => {
+    if (err) {
+      console.error(err.message);
+    }
+    console.log('Connected to the chinook database.');
+  });
+
+  return db;
+};
+
+const insertFinData = (db, data, callback) => {
+  let sql = '', id = '';
+  if (Array.isArray(data)) {
+    data.map(({id, category, subcategory, money, comment, date}) => {
+      let tmpSql =
+        `insert into ${TABLE_NAME}(id, category, subcategory, date, comment, money)
+        values ("${id}", "${category}", "${subcategory}", "${date}", "${comment}", "${money}");`;
+      sql += tmpSql;
+    });
+  } else {
+    id = data.id;
+    sql =
+      `insert into ${TABLE_NAME}(id, category, subcategory, date, comment, money)
+      values ("${data.id}", "${data.category}", "${data.subcategory}", "${data.date}", "${data.comment}", "${data.money}");`;
+  }
+
+  db.run(sql, (err) => {
+    if (err) {
+      logDBError('insert data with id: ' + id, err);
+    } else {
+      logDBSuccess('insert data with id: ' + id);
+    }
+
+    callback && callback(err);
+  });
+};
+
+const deleteAllData = (db, callback) => {
+  let sql = `delete from ${TABLE_NAME};`;
+  db.run(sql, (err) => {
+    if (err) {
+      logDBError('Delete all data in FIN table', err);
+    } else {
+      logDBSuccess('Delete all data in FIN table');
+    }
+
+    callback && callback(err);
+  });
+}
+
+const closeDB = (db) => {
+  db.close();
+};
+
+module.exports = {
+  createDBConnection,
+  insertFinData,
+  deleteAllData,
+  closeDB,
+};
+